Struct customsearch1::CseListMethodBuilder

pub struct CseListMethodBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

Returns metadata about the search performed, metadata about the custom search engine used for the search, and the search results.

A builder for the list method supported by a cse resource. It is not used directly, but through a CseMethodsBuilder.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.cse().list("q")
             .start(6)
             .sort("rebum.")
             .site_search_filter("dolore")
             .site_search("nonumy")
             .search_type("sed")
             .safe("aliquyam")
             .rights("sit")
             .related_site("eirmod")
             .or_terms("consetetur")
             .num(16)
             .lr("sed")
             .low_range("ea")
             .link_site("gubergren")
             .img_type("aliquyam")
             .img_size("eos")
             .img_dominant_color("tempor")
             .img_color_type("sea")
             .hq("labore")
             .hl("ipsum")
             .high_range("aliquyam")
             .googlehost("dolores")
             .gl("sit")
             .filter("diam")
             .file_type("ut")
             .exclude_terms("justo")
             .exact_terms("est")
             .date_restrict("amet")
             .cx("accusam")
             .cref("clita")
             .cr("diam")
             .c2coff("justo")
             .doit();

Methods

impl<'a, C, NC, A> CseListMethodBuilder<'a, C, NC, A> where NC: NetworkConnector, C: BorrowMut<Client<NC>> + 'a, A: GetToken

fn doit(self) -> Result<Search>

Perform the operation you have build so far.

fn q(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the q query property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

Query

fn start(self, new_value: u32) -> CseListMethodBuilder<'a, C, NC, A>

Sets the start query property to the given value.

The index of the first result to return

fn sort(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the sort query property to the given value.

The sort expression to apply to the results

fn site_search_filter(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the site search filter query property to the given value.

Controls whether to include or exclude results from the site named in the as_sitesearch parameter

fn site_search(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the site search query property to the given value.

Specifies all search results should be pages from a given site

fn search_type(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the search type query property to the given value.

Specifies the search type: image.

fn safe(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the safe query property to the given value.

Search safety level

fn rights(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the rights query property to the given value.

Filters based on licensing. Supported values include: cc_publicdomain, cc_attribute, cc_sharealike, cc_noncommercial, cc_nonderived and combinations of these.

fn related_site(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the related site query property to the given value.

Specifies that all search results should be pages that are related to the specified URL

fn or_terms(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the or terms query property to the given value.

Provides additional search terms to check for in a document, where each document in the search results must contain at least one of the additional search terms

fn num(self, new_value: u32) -> CseListMethodBuilder<'a, C, NC, A>

Sets the num query property to the given value.

Number of search results to return

fn lr(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the lr query property to the given value.

The language restriction for the search results

fn low_range(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the low range query property to the given value.

Creates a range in form as_nlo value..as_nhi value and attempts to append it to query

fn link_site(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the link site query property to the given value.

Specifies that all search results should contain a link to a particular URL

fn img_type(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the img type query property to the given value.

Returns images of a type, which can be one of: clipart, face, lineart, news, and photo.

fn img_size(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the img size query property to the given value.

Returns images of a specified size, where size can be one of: icon, small, medium, large, xlarge, xxlarge, and huge.

fn img_dominant_color(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the img dominant color query property to the given value.

Returns images of a specific dominant color: yellow, green, teal, blue, purple, pink, white, gray, black and brown.

fn img_color_type(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the img color type query property to the given value.

Returns black and white, grayscale, or color images: mono, gray, and color.

fn hq(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the hq query property to the given value.

Appends the extra query terms to the query.

fn hl(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the hl query property to the given value.

Sets the user interface language.

fn high_range(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the high range query property to the given value.

Creates a range in form as_nlo value..as_nhi value and attempts to append it to query

fn googlehost(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the googlehost query property to the given value.

The local Google domain to use to perform the search.

fn gl(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the gl query property to the given value.

Geolocation of end user.

fn filter(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the filter query property to the given value.

Controls turning on or off the duplicate content filter.

fn file_type(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the file type query property to the given value.

Returns images of a specified type. Some of the allowed values are: bmp, gif, png, jpg, svg, pdf, ...

fn exclude_terms(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the exclude terms query property to the given value.

Identifies a word or phrase that should not appear in any documents in the search results

fn exact_terms(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the exact terms query property to the given value.

Identifies a phrase that all documents in the search results must contain

fn date_restrict(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the date restrict query property to the given value.

Specifies all search results are from a time period

fn cx(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the cx query property to the given value.

The custom search engine ID to scope this search query

fn cref(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the cref query property to the given value.

The URL of a linked custom search engine

fn cr(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the cr query property to the given value.

Country restrict(s).

fn c2coff(self, new_value: &str) -> CseListMethodBuilder<'a, C, NC, A>

Sets the c2coff query property to the given value.

Turns off the translation between zh-CN and zh-TW.

fn delegate(self, new_value: &'a mut Delegate) -> CseListMethodBuilder<'a, C, NC, A>

Sets the delegate property to the given value.

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

fn param<T>(self, name: T, value: T) -> CseListMethodBuilder<'a, C, NC, A> where T: Str

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.