Struct customsearch1::Search

pub struct Search {
    pub promotions: Vec<Promotion>,
    pub kind: Option<String>,
    pub url: Option<SearchUrl>,
    pub items: Vec<Result>,
    pub context: Option<Context>,
    pub queries: HashMap<String, Vec<Query>>,
    pub spelling: Option<SearchSpelling>,
    pub search_information: Option<SearchSearchInformation>,
}

There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list cse (response)

Fields

promotions

no description provided

kind

no description provided

url

no description provided

items

no description provided

context

no description provided

queries

no description provided

spelling

no description provided

search_information

no description provided